+++ config/config/src/main/java/io/helidon/config/InMemoryConfigSource.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2019 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2017, 2020 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,76 +19,84 @@ package io.helidon.config;
 import java.util.Objects;
 import java.util.Optional;
 
-import io.helidon.config.spi.AbstractParsableConfigSource;
+import io.helidon.config.spi.ConfigContent.NodeContent;
 import io.helidon.config.spi.ConfigParser;
+import io.helidon.config.spi.ConfigSource;
+import io.helidon.config.spi.NodeConfigSource;
+import io.helidon.config.spi.ParsableSource;
 
 /**
  * In-memory implementation of config source.
  */
-class InMemoryConfigSource extends AbstractParsableConfigSource<Object> {
-
-    private final String uri;
-    private final ConfigParser.Content<Object> content;
-
-    InMemoryConfigSource(InMemoryConfigSource.Builder builder) {
-        super(builder);
-
-        uri = builder.uri();
-        content = builder.content();
+class InMemoryConfigSource {
+    private InMemoryConfigSource() {
     }
 
-    @Override
-    protected String uid() {
-        return uri;
+    static NodeConfigSource create(String uri, NodeContent content) {
+        Objects.requireNonNull(uri, "uri cannot be null");
+        Objects.requireNonNull(content, "content cannot be null");
+
+        return new NodeInMemory(uri, content);
     }
 
-    @Override
-    protected Optional<Object> dataStamp() {
-        return Optional.of(this);
+    static ConfigSource create(String uri, ConfigParser.Content content) {
+        Objects.requireNonNull(uri, "uri cannot be null");
+        Objects.requireNonNull(content, "content cannot be null");
+
+        return new ParsableInMemory(uri, content);
     }
 
-    @Override
-    protected ConfigParser.Content<Object> content() throws ConfigException {
-        return content;
-    }
+    private static class InMemory implements ConfigSource {
+        private final String uid;
 
-    static Builder builder() {
-        return new Builder();
-    }
-
-    static final class Builder
-            extends AbstractParsableConfigSource.Builder<InMemoryConfigSource.Builder, Void, InMemoryConfigSource> {
-
-        private String uri;
-        private ConfigParser.Content<Object> content;
-
-        Builder() {
-            super(Void.class);
-        }
-
-        Builder content(String uri, ConfigParser.Content<Object> content) {
-            Objects.requireNonNull(uri, "uri cannot be null");
-            Objects.requireNonNull(content, "content cannot be null");
-
-            this.uri = uri;
-            this.content = content;
-            return this;
+        protected InMemory(String uid) {
+            this.uid = uid;
         }
 
         @Override
-        public InMemoryConfigSource build() {
-            Objects.requireNonNull(uri, "uri cannot be null");
-            Objects.requireNonNull(content, "content cannot be null");
+        public String description() {
+            return ConfigSource.super.description()
+                    + "[" + uid + "]";
+        }
+    }
 
-            return new InMemoryConfigSource(this);
+    private static final class NodeInMemory extends InMemory implements NodeConfigSource {
+        private final NodeContent content;
+
+        private NodeInMemory(String uid,  NodeContent nodeContent) {
+            super(uid);
+            this.content = nodeContent;
         }
 
-        private String uri() {
-            return uri;
+        @Override
+        public Optional<NodeContent> load() throws ConfigException {
+            return Optional.of(content);
+        }
+    }
+
+    private static final class ParsableInMemory extends InMemory implements ParsableSource {
+        private final ConfigParser.Content content;
+
+        protected ParsableInMemory(String uid, ConfigParser.Content content) {
+            super(uid);
+            this.content = content;
         }
 
-        private ConfigParser.Content<Object> content() {
-            return content;
+        @Override
+        public Optional<ConfigParser.Content> load() throws ConfigException {
+            return Optional.of(content);
+        }
+
+        @Override
+        public Optional<ConfigParser> parser() {
+            return Optional.empty();
+        }
+
+        @Override
+        public Optional<String> mediaType() {
+            return Optional.empty();
         }
     }
 }
+
+