+++ config/config/src/test/java/io/helidon/config/ConfigContextTest.java
@@ -23,7 +23,6 @@ import java.util.concurrent.TimeUnit;
 import java.util.stream.Stream;
 
 import io.helidon.config.spi.ConfigNode.ObjectNode;
-import io.helidon.config.spi.TestingConfigSource;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -42,25 +41,28 @@ public class ConfigContextTest {
     private static final String PROP1 = "prop1";
     private static final int TEST_DELAY_MS = 1;
 
-    private class TestContext {
-        
+    private static class TestContext {
+
         private final Config config;
         private final String key;
         private final String oldValue;
         private final String newValue;
         private final TestingConfigSource configSource;
 
-        
-        private TestContext(Map.Entry<String,String> entry) {
+        private TestContext(Map.Entry<String, String> entry) {
             String detachKey = entry.getKey();
             key = entry.getValue();
 
-            int i = detachKey == null || detachKey.length() == 0 ? 1 : 2;
-            int j = key.length() == 0 ? 1 : 2;
+            int i = ((detachKey == null) || detachKey.isEmpty()) ? 1 : 2;
+            int j = key.isEmpty() ? 1 : 2;
             oldValue = "oldVal_" + i + "_" + j;
             newValue = "newVal_" + i + "_" + j;
 
-            configSource = TestingConfigSource.builder().objectNode(createSource("old")).build();
+            configSource = TestingConfigSource.builder()
+                    .objectNode(createSource("old"))
+                    .testingPollingStrategy()
+                    .build();
+
             Config cfg = Config.builder()
                     .sources(configSource)
                     .disableEnvironmentVariablesSource()
@@ -72,10 +74,10 @@ public class ConfigContextTest {
             }
             config = cfg.get(key);
         }
-        
-        private void changeSource(boolean submitChange, String valuePrefix) throws InterruptedException {
+
+        private void changeSource(String valuePrefix) throws InterruptedException {
             TimeUnit.MILLISECONDS.sleep(TEST_DELAY_MS); // Make sure timestamp will change.
-            configSource.changeLoadedObjectNode(createSource(valuePrefix), submitChange);
+            configSource.changeLoadedObjectNode(createSource(valuePrefix));
         }
     }
     
@@ -106,18 +108,10 @@ public class ConfigContextTest {
     }
 
     private static String concatKeys(String prefix, String suffix) {
-        if (prefix.length() == 0) {
-            if (suffix.length() == 0) {
-                return "";
-            } else {
-                return suffix;
-            }
+        if (prefix.isEmpty()) {
+            return suffix;
         } else {
-            if (suffix.length() == 0) {
-                return prefix;
-            } else {
-                return prefix + "." + suffix;
-            }
+            return suffix.isEmpty() ? prefix : (prefix + "." + suffix);
         }
     }
 
@@ -165,10 +159,12 @@ public class ConfigContextTest {
         assertThat(c.config.get(PROP1).asString().get(), is(c.oldValue));
 
         // CHANGE source
-        c.changeSource(false, "new");
+        c.changeSource("new");
 
         // RELOAD config
         TimeUnit.MILLISECONDS.sleep(TEST_DELAY_MS); // Make sure the timestamp changes.
+        // wait until the change is propagated
+        listener.get(500, true);
         Config reloaded = c.config.context().reload();
 
         // new config -> new timestamp
@@ -198,7 +194,7 @@ public class ConfigContextTest {
 
         // CHANGE source
         TimeUnit.MILLISECONDS.sleep(TEST_DELAY_MS); // Make sure timestamp changes.
-        c.changeSource(true, "old");
+        c.changeSource("old");
 
         // no other events
         assertThat(listener.get(500, false), is(nullValue()));
@@ -219,7 +215,7 @@ public class ConfigContextTest {
 
         // CHANGE source
         TimeUnit.MILLISECONDS.sleep(TEST_DELAY_MS); // Make sure time changes to trigger notification.
-        c.changeSource(true, "new");
+        c.changeSource("new");
 
         // change event
         Config last1 = listener.get(500, true);
@@ -243,10 +239,10 @@ public class ConfigContextTest {
 
         // CHANGE source
         TimeUnit.MILLISECONDS.sleep(TEST_DELAY_MS); // Make sure time changes to trigger notification.
-        c.changeSource(true, "new");
+        c.changeSource("new");
 
         //wait for a new configuration is loaded
-        waitForAssert(() -> c.config.context().last().get(PROP1).asString(), is(ConfigValues.simpleValue(c.newValue)));
+        waitForAssert(() -> c.config.context().last().get(PROP1).asString().asOptional(), is(ConfigValues.simpleValue(c.newValue).asOptional()));
 
         Config last1 = c.config.context().last();