+++ config/config/src/test/java/io/helidon/config/ConfigHelperTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2019 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2017, 2020 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,16 +16,11 @@
 
 package io.helidon.config;
 
-import java.io.IOException;
-import java.io.StringReader;
-import java.nio.CharBuffer;
 import java.util.concurrent.Flow;
 import java.util.function.Function;
 
 import org.junit.jupiter.api.Test;
 
-import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
@@ -37,19 +32,6 @@ import static org.mockito.Mockito.when;
  * Tests {@link ConfigHelper}.
  */
 public class ConfigHelperTest {
-
-    @Test
-    public void testCreateReaderFromReader() throws IOException {
-        StringReader readable = new StringReader("test-value");
-
-        assertThat(ConfigHelper.createReader(readable), is(readable));
-    }
-
-    @Test
-    public void testCreateReaderFromCharBuffer() throws IOException {
-        assertThat(readerToString(ConfigHelper.createReader(CharBuffer.wrap("test-value"))), is("test-value"));
-    }
-
     @Test
     public void testSubscriber() {
         //mocks
@@ -79,26 +61,6 @@ public class ConfigHelperTest {
         //    function invoked 2+1x, cancel 1x
         verify(onNextFunction, times(2 + 1)).apply(any());
         verify(subscription, times(1)).cancel();
-    }
-
-    /**
-     * Reads specified {@code readable} into String.
-     *
-     * @param readable readable
-     * @return String
-     * @throws IOException in case of error during reading from readable
-     */
-    public static String readerToString(Readable readable) throws IOException {
-        StringBuilder builder = new StringBuilder();
-        CharBuffer chars = CharBuffer.allocate(100);
-        int charsRead;
-        do {
-            charsRead = readable.read(chars);
-            if (charsRead > 0) {
-                builder.append(chars.flip().toString());
-            }
-        } while (charsRead > 0);
-        return builder.toString();
     }
 
 }